+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 * Added modifiable arena state names to language
 * Added configurable map selector items
 * Added legacy support
+* Added support for multiple last words that will randomize
+* Added a lot of new Trails (Should support every particle name that is supported since 1.8)
+* Changed trails will now randomize if you got perms for more than one  
+* Cached armor stands will be removed after block break if type is matches
+* Now special block holograms will get removed after game stop
+  * After game start, it will show again
 
 ### 1.7.4 Release (25.01.2021 - 28.01.2021)
 * Fix stack trace array length is 0 when trying to send exception to service